requirements.txt
nodriver>=0.34
colorama>=0.4.6
python-dotenv>=1.0.0
requests>=2.31.0
